--- a/pkg/database/handler.go
+++ b/pkg/database/handler.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	// the injected postgres interface implementations for Go SQL
-	_ "github.com/lib/pq"
+	"github.com/lib/pq"
 
 	"github.com/open-sauced/pizza/oven/pkg/insights"
 )
@@ -65,23 +65,124 @@ func (p PizzaOvenDbHandler) GetAuthorID(insight insights.CommitInsight) (int, er
 	return id, err
 }
 
-// InsertAuthor inserts an author by their email
-func (p PizzaOvenDbHandler) InsertAuthor(insight insights.CommitInsight) (int, error) {
-	var id int
-	err := p.db.QueryRow("INSERT INTO public.commit_authors(commit_author_email) VALUES($1) RETURNING id", insight.AuthorEmail).Scan(&id)
-	return id, err
+// GetAuthorIDs queries the id of an author by their email
+func (p PizzaOvenDbHandler) GetAuthorIDs(emails []string) (map[string]int, error) {
+	emailIDMap := make(map[string]int)
+
+	rows, err := p.db.Query("SELECT id, commit_author_email FROM commit_authors WHERE commit_author_email = ANY($1);", pq.Array(emails))
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer rows.Close()
+	for rows.Next() {
+		var id int
+		var email string
+		if err := rows.Scan(&id, &email); err != nil {
+			log.Fatal(err)
+		}
+		emailIDMap[email] = id
+	}
+
+	return emailIDMap, nil
 }
 
-// GetCommitID queries the id of a given commit based on its hash
-func (p PizzaOvenDbHandler) GetCommitID(repoID int, insight insights.CommitInsight) (int, error) {
-	var id int
-	err := p.db.QueryRow("SELECT id FROM public.commits WHERE baked_repo_id=$1 AND commit_hash=$2", repoID, insight.Hash).Scan(&id)
-	return id, err
+// PrepareBulkAuthorInsert creates a temporary table that mirrors the commit_authors
+// and is used to perform a bulk insert "pivot" which accounts for conflicts
+func (p PizzaOvenDbHandler) PrepareBulkAuthorInsert(tmpTableName string) (*sql.Tx, *sql.Stmt, error) {
+	_, err := p.db.Exec(fmt.Sprintf("CREATE TEMPORARY TABLE %s AS SELECT * FROM commit_authors WHERE 1=0", tmpTableName))
+	if err != nil {
+		return nil, nil, err
+	}
+
+	txn, err := p.db.Begin()
+	if err != nil {
+		return nil, nil, err
+	}
+
+	stmt, err := txn.Prepare(pq.CopyIn(tmpTableName, "commit_author_email"))
+	if err != nil {
+		newErr := txn.Rollback()
+		if newErr != nil {
+			return nil, nil, fmt.Errorf("could not abort the sql transaction: %s - original error: %s", newErr, err)
+		}
+
+		return nil, nil, err
+	}
+
+	return txn, stmt, nil
 }
 
-// InsertCommit inserts a commit based on its commit hash
-func (p PizzaOvenDbHandler) InsertCommit(insight insights.CommitInsight, authorID int, repoID int) error {
-	_, err := p.db.Exec("INSERT INTO public.commits(commit_hash, commit_author_id, baked_repo_id, commit_date) VALUES($1, $2, $3, $4)", insight.Hash, authorID, repoID, insight.Date)
+// PivotTmpTableToAuthorsTable performs the pivot from the temporary commit authors
+// table to the real one handling any conflicts
+func (p PizzaOvenDbHandler) PivotTmpTableToAuthorsTable(tmpTableName string) error {
+	_, err := p.db.Exec(fmt.Sprintf(`
+		INSERT INTO public.commit_authors(commit_author_email)
+		SELECT commit_author_email FROM %s
+		ON CONFLICT (commit_author_email)
+		DO NOTHING
+	`, tmpTableName))
+	if err != nil {
+		return err
+	}
+
+	_, err = p.db.Exec(fmt.Sprintf("DROP TABLE %s", tmpTableName))
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// InsertAuthor inserts an author by their email into the sql transaction
+func (p PizzaOvenDbHandler) InsertAuthor(stmt *sql.Stmt, insight insights.CommitInsight) error {
+	_, err := stmt.Exec(insight.AuthorEmail)
+	return err
+}
+
+// PrepareBulkCommitInsert gets a sql bulk transaction ready to insert all commits
+// from processing in one round trip
+func (p PizzaOvenDbHandler) PrepareBulkCommitInsert() (*sql.Tx, *sql.Stmt, error) {
+	txn, err := p.db.Begin()
+	if err != nil {
+		return nil, nil, err
+	}
+
+	stmt, err := txn.Prepare(pq.CopyIn("commits", "commit_hash", "commit_author_id", "baked_repo_id", "commit_date"))
+	if err != nil {
+		newErr := txn.Rollback()
+		if newErr != nil {
+			return nil, nil, fmt.Errorf("could not abort commits bulk sql transaction: %s - original error: %s", newErr, err)
+		}
+
+		return nil, nil, err
+	}
+
+	return txn, stmt, nil
+}
+
+// ResolveTransaction resolves a given transaction and sql statement
+func (p PizzaOvenDbHandler) ResolveTransaction(txn *sql.Tx, stmt *sql.Stmt) error {
+	_, err := stmt.Exec()
+	if err != nil {
+		return err
+	}
+
+	err = stmt.Close()
+	if err != nil {
+		return err
+	}
+
+	err = txn.Commit()
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+// InsertCommit adds a commit to the given sql.Stmt to be executed in bulk
+func (p PizzaOvenDbHandler) InsertCommit(stmt *sql.Stmt, insight insights.CommitInsight, authorID int, repoID int) error {
+	_, err := stmt.Exec(insight.Hash, authorID, repoID, insight.Date)
 	return err
 }