## Description

This optimizes the database calls within the pizza oven.

Instead of iterating and inserting _each_ author and commit, this uses methods to bulk insert those records:
- Each author is added to a transaction. When that transaction is executed and if there is a conflict, the transaction for that record "does nothing". This takes 1 call to the database.
  - The commit authors are re-queried in order to build a map of inserted authors emails and IDs. These IDs are used when bulk inserting into the commits table.
- Each commit is added to a transaction. When the transaction is executed, each commit is inserted.

On some preliminary testing using the freecodecamp/freecodecamp repo (which as about ~30,000 commits), each with a fresh local database (so network latency should be negligible):

Using `main`:
```
❯ time pizza bake -e http://localhost:8080 https://github.com/freecodecamp/freecodecamp --wait
Resp body:
pizza bake -e http://localhost:8080  --wait  0.01s user 0.02s system 0% cpu 1:17.42 total
```

Using this branch:
```
❯ time pizza bake -e http://localhost:8080 https://github.com/freecodecamp/freecodecamp --wait
Resp body:
pizza bake -e http://localhost:8080  --wait  0.01s user 0.02s system 0% cpu 3.322 total
```

which ends up being about a 24x faster.

## What type of PR is this? (check all applicable)

- [x] 🍕 Feature
- [ ] 🐛 Bug Fix
- [ ] 📝 Documentation Update
- [ ] 🎨 Style
- [ ] 🧑‍💻 Code Refactor
- [ ] 🔥 Performance Improvements
- [ ] ✅ Test
- [ ] 🤖 Build
- [ ] 🔁 CI
- [ ] 📦 Chore (Release)
- [ ] ⏩ Revert

## Related Tickets & Documents

N/a - related to doing some scaling and deployments after integrating with some of our internal services.

## Mobile & Desktop Screenshots/Recordings

N/a

## Added tests?

- [ ] 👍 yes
- [ ] 🙅 no, because they aren't needed
- [ ] 🙋 no, because I need help

## Added to documentation?

- [ ] 📜 README.md
- [ ] 📓 docs.opensauced.pizza
- [ ] 🍕 dev.to/opensauced
- [ ] 📕 storybook
- [ ] 🙅 no documentation needed

## [optional] Are there any post-deployment tasks we need to perform?



## [optional] What gif best describes this PR or how it makes you feel?



<!-- note: PRs with deleted sections will be marked invalid -->

<!--
  For Work In Progress Pull Requests, please use the Draft PR feature,
  see https://github.blog/2019-02-14-introducing-draft-pull-requests/ for further details.
  
  For a timely review/response, please avoid force-pushing additional
  commits if your PR already received reviews or comments.
  
  Before submitting a Pull Request, please ensure you've done the following:
  - 📖 Read the Open Sauced Contributing Guide: https://github.com/open-sauced/.github/blob/main/CONTRIBUTING.md.
  - 📖 Read the Open Sauced Code of Conduct: https://github.com/open-sauced/.github/blob/main/CODE_OF_CONDUCT.md.
  - 👷‍♀️ Create small PRs. In most cases, this will be possible.
  - ✅ Provide tests for your changes.
  - 📝 Use descriptive commit messages.
  - 📗 Update any related documentation and include any relevant screenshots.
-->